service/src/test/java/backend/server/service/Service/CompagnieServiceTest.java
package backend.server.service.Service;

import backend.server.service.Repository.CompagnieRepository;
import backend.server.service.domain.Compagnie;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.Mockito;
import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;

class CompagnieServiceTest {
    @Mock
    private CompagnieRepository compagnieRepository;
    @InjectMocks
    private CompagnieService compagnieService;
    @Test
    void getCompagnie() {
    }
    @Test
    void testGetCompagnie() {
    }

    @Test
    void getAllCompagnies() {
    }

    @Test
    void testcreateCompagnie() {
        // Create a sample compagnie object
        Compagnie compagnie = new Compagnie();
        compagnie.setId(1L);
        compagnie.setNom("Sample Company");

        // Mock the save operation of the compagnie repository
        when(compagnieRepository.save(Mockito.any(Compagnie.class))).thenReturn(compagnie);

        // Call the createCompagnie method
        Compagnie result = compagnieService.createCompagnie(compagnie);

        // Verify that the save operation was called with the correct parameter
        verify(compagnieRepository).save(compagnie);

        // Assert the result
        assertEquals(compagnie.getId(), result.getId());
        assertEquals(compagnie.getNom(), result.getNom());
    }

    @Test
    void updateCompagnie() {
    }

    @Test
    void deleteCompagnie() {
    }

    @Test
    void testDeleteCompagnie() {
    }

    @Test
    void createGroupe() {
    }

    @Test
    void testCreateGroupe() {
    }

    @Test
    void deleteGroupe() {
    }

    @Test
    void updateGroupe() {
    }

    @Test
    void deleteMembre() {
    }

    @Test
    void updateMembre() {
    }

    @Test
    void getAllUniqueGroups() {
    }
}